requirements-test.txt
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov
